Moved apod-cache to /tmp instead of project dir

+++ apodutils.js
@@ -4,7 +4,8 @@ var fetch = require("node-fetch");
 
 const API_KEY = fs.readFileSync('API_KEY', { encoding: 'utf-8', flag: 'r' });
 const APOD_URL = `https://api.nasa.gov/planetary/apod`
-const ARCHIVE_DIR = __dirname + '/archive/'
+const ARCHIVE_DIR = '/tmp/apod-cache/'
+if (!fs.existsSync(ARCHIVE_DIR)) fs.mkdirSync(ARCHIVE_DIR)
 
 function formatDate(date) {
   return new Date(date).toISOString().split('T')[0]
@@ -13,14 +14,16 @@ function formatDate(date) {
 function cacheAPOD(apod) {
 	apodString = JSON.stringify(apod, null, 2)
 	filename = ARCHIVE_DIR + apod.date + '.json'
-	fs.writeFileSync(filename, apodString)
+	if (!fs.existsSync(filename)) fs.writeFileSync(filename, apodString)
+	return filename
 }
 
 async function getAPOD(date=new Date()) {
   let dateISO = formatDate(date)
   let apod = {}
-  if (fs.existsSync(ARCHIVE_DIR + date)) {
-  	let apodString = fs.readFileSync(ARCHIVE_DIR + date,
+  console.log(dateISO)
+  if (fs.existsSync(ARCHIVE_DIR + dateISO + '.json')) {
+  	let apodString = fs.readFileSync(ARCHIVE_DIR + dateISO + '.json',
   									{ encoding: 'utf-8', flag: 'r' })
   	apod = JSON.parse(apodString)
   } else {
